backend/requirements.txt
Django==4.1
djangorestframework==3.14.0
joblib==1.1.1
numpy==1.23.5
pandas==1.5.2
scikit_learn==1.2.0
xgboost
gunicorn
